refactor(shared): wrap summarize_text in summarize_report
