+++ .gitignore
@@ -25,3 +25,4 @@ parsed_data/
 Thumbs.db
 
 repo-structure.txt
+clean_tree.txt
