+++ functions/cleaner_function/function_app.py
@@ -29,7 +29,7 @@ from tenacity import retry, wait_exponential, stop_after_attempt, retry_if_excep
 
 from backend.api.transcription import transcribe_audio
 from backend.api.summarization import summarize_text
-from backend.api.form_cleaner import clean_form_data
+from backend.shared.form_cleaner import clean_form_data
 from backend.utils.validation import (
     validate_medical_data,
     remove_empty_elements,
